Fixed better-auth for User and Admin for all files in the project where validation is needed

+++ src/actions/person.ts
@@ -1,4 +1,5 @@
 "use server";
+import { requireAdmin } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import {
   createPersonSchema,
@@ -11,6 +12,10 @@ import {
 // !TODO auth checks!
 // !!!!!!!!!!!!!!!!!!
 export async function createPerson(person: CreatePersonInput) {
+
+   //Authorization
+    await requireAdmin();
+    
   const data = await createPersonSchema.parseAsync(person);
 
   const newPerson = await prisma.person.create({
@@ -28,6 +33,10 @@ export async function createPerson(person: CreatePersonInput) {
 }
 
 export async function updatePerson(person: UpdatePersonInput) {
+
+   //Authorization
+  await requireAdmin();
+
   const data = await updatePersonSchema.parseAsync(person);
 
   const updatedPerson = await prisma.person.update({
@@ -46,6 +55,10 @@ export async function updatePerson(person: UpdatePersonInput) {
 }
 
 export async function deletePerson(id: number) {
+
+   //Authorization
+  await requireAdmin();
+  
   const deletedPerson = await prisma.person.delete({
     where: { id },
   });
